% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTRee.R
\name{bootstrap_collapse}
\alias{bootstrap_collapse}
\title{bootstrap_collapse}
\usage{
bootstrap_collapse(tree, cutoff = 50)
}
\arguments{
\item{tree}{An object representing the phylogenetic tree. Should be of class 'treedata' or 'phylo'.}

\item{cutoff}{The threshold value for collapsing nodes based on bootstrap support. Default: 50}
}
\value{
A treedata object with collapsed branches where bootstrap support is less than the user-specified cutoff.
}
\description{
Collapse nodes in a phylogenetic tree where the bootstrap support is below a user-specified cutoff.
If bootstrap support is in 0-1 scale, is converted at the 1-100 scale.
}
\examples{
\dontrun{
  # Load a sample tree
  tree <- read_tree("path/to/tree.nwk")

  # Collapse nodes with bootstrap support less than 0.7
  collapsed_tree <- bootstrap_collapse(tree, cutoff = 0.7)
}

}
