% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R, R/OliveTree_CMD.R
\name{.load_tree_object}
\alias{.load_tree_object}
\title{.load_tree_object
Load phylogenetic tree file and/or object as treedata object}
\usage{
.load_tree_object(tree)

.load_tree_object(tree)
}
\arguments{
\item{tree}{An object representing the phylogenetic tree. Should be a newick or nexus file, or an object of class 'treedata' or 'phylo'.}
}
\description{
.load_tree_object
Load phylogenetic tree file and/or object as treedata object

.load_tree_object
Load phylogenetic tree file and/or object as treedata object
}
