% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{export_plot}
\alias{export_plot}
\title{export_plot: Function to export plot with provided options}
\usage{
export_plot(plot, save = TRUE, output = "tree_plot_visualized.svg")
}
\arguments{
\item{plot}{A ggtree plot object.}

\item{save}{Option to save plot to output svg file. Default: True}

\item{output}{Name of the output file. Default: 'tree_plot_visualized.svg'}
}
\value{
A ggplot object representing the phylogenetic tree and optionally an output svg file with this phylogenetic tree.
Main plot exporting function of the package.
}
\description{
export_plot: Function to export plot with provided options
}
\examples{
dontrun{
# Export plot with option to save
export_plot (plot, save = TRUE, output = "tree_plot_visualized.svg")
}
}
