% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{flip_nodes}
\alias{flip_nodes}
\title{flip_nodes: Flip nodes on a phylogenetic tree based on the specified descendant nodes. It can flip internal nodes or leaves, if the node is terminal.}
\usage{
flip_nodes(tree, node1, node2)
}
\arguments{
\item{tree}{An object representing the phylogenetic tree. Should be of class 'treedata' or 'phylo'.}

\item{node1}{The first node for flipping.}

\item{node2}{The second node for flipping.}
}
\value{
A 'phylo' object with flipped nodes.
}
\description{
flip_nodes: Flip nodes on a phylogenetic tree based on the specified descendant nodes. It can flip internal nodes or leaves, if the node is terminal.
}
\examples{
\dontrun{
# Load a sample tree
tree <- read_tree("path/to/tree.nwk")

# Use node_ids to identify the nodes you want to flip, ideally with a pattern matching a leaf of interest.
node_ids(tree)

# Flip nodes 5 and 8 in the tree
flipped_tree <- flip_node(tree, 5, 8)
}

}
