% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{group_descendants}
\alias{group_descendants}
\title{group_descendants: Group all descendant branches of specified node(s) in a phylogenetic tree.
This function groups all descendant branches of the specified node(s) in a phylogenetic tree.}
\usage{
group_descendants(tree, ...)
}
\arguments{
\item{tree}{An object representing the phylogenetic tree. Should be of class 'treedata' or 'phylo'.}

\item{node1}{The first node for grouping descendants.}

\item{node2}{The second node for grouping descendants.}

\item{node3}{The third node for grouping descendants.}

\item{node4}{The fourth node for grouping descendants.}
}
\value{
A 'phylo' object with grouped descendants.
}
\description{
group_descendants: Group all descendant branches of specified node(s) in a phylogenetic tree.
This function groups all descendant branches of the specified node(s) in a phylogenetic tree.
}
\examples{
\dontrun{
# Load a sample tree
tree <- read_tree("path/to/tree.nwk")

# Group all descendants of nodes 5, 8, and 10 in the tree
grouped_tree <- group_descendants(tree, 5, 8, 10)
}

}
