% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{highlight_tree}
\alias{highlight_tree}
\title{highlight_tree: Highlight Nodes and Descendant Leaves on a Phylogenetic Tree}
\usage{
highlight_tree(
  tree,
  highlight_nodes,
  legend_position = "right",
  legend_text_position = "right",
  legend_orientation = "horizontal",
  legend_key_size = 0.1,
  legend_font = "Arial",
  legend_fontsize = 11,
  legend_font_face = "bold",
  legend_spacing_x = 0.5,
  legend_spacing_y = 0.5,
  legend_key_width = 1,
  legend_title_hjust = 0.5,
  tip_fill_color = "lightgrey",
  tip_border_color = "white",
  colors = NULL,
  form = "circular",
  legend_name = "",
  name = NULL,
  save = TRUE,
  output = "highlighted_tree.svg"
)
}
\arguments{
\item{tree}{A phylogenetic tree object, typically of class 'phylo' from the ape package,
or a file path to a tree file in Newick format that can be read into such an object.}

\item{highlight_nodes}{An associative vector where names are group labels and values are
node IDs indicating the nodes and their descendants to be highlighted on the tree.}

\item{legend_position}{Position of the legend in the plot ("right", "left", "top", "bottom").}

\item{legend_text_position}{Horizontal alignment of the text within the legend ("left" or "right").}

\item{legend_orientation}{Orientation of the legend ("horizontal" or "vertical").}

\item{legend_key_size}{Size of the legend keys, specified in cm.}

\item{legend_font}{Font family for the text in the legend.}

\item{legend_fontsize}{Size of the text in the legend, specified in points.}

\item{legend_font_face}{Boolean indicating whether the text in the legend should be bold.}

\item{legend_spacing_x}{Horizontal spacing between legend items, specified in cm.}

\item{legend_spacing_y}{Vertical spacing between legend items, specified in cm.}

\item{legend_key_width}{Width of the keys in the legend, specified in cm.}

\item{legend_title_hjust}{Horizontal justification of the legend title (0 to 1).}

\item{tip_fill_color}{Fill color for the tip label symbols. Default: lightgrey}

\item{tip_border_color}{Border color for the tip label symbols. Default: black}

\item{colors}{An associative vector where names are group labels and values are colors
for the corresponding groups defined in \code{highlight_nodes}. If not provided,
random colors will be assigned to each group.}

\item{form}{A character string specifying the layout of the tree. Common values include
"rectangular", "circular", and others supported by ggtree.}

\item{legend_name}{A character string specifying the legend name for the highlighted plot. This parameter
is optional and primarily used for labeling within complex workflows.}

\item{save}{Logical; if TRUE, the plot will be saved to the file specified in \code{output}.}

\item{output}{The filename where the plot will be saved if \code{save} is TRUE.}
}
\value{
A ggplot object representing the phylogenetic tree with highlighted nodes.
}
\description{
This function highlights specified nodes and their descendant leaves on a phylogenetic tree,
allowing for visualization emphasis on selected parts of the tree. Custom colors can be assigned
to different groups of nodes, or random colors can be automatically generated.
}
\examples{
\dontrun{
library(ggtree)
library(treeio)
# Load a tree from a Newick file
tree <- read.tree("path/to/your/treefile.newick")
# Highlight specific nodes with specified colors
highlight_tree(tree,
               highlight_nodes = c("Node1" = 5, "Node2" = 10),
               colors = c("Node1" = "red", "Node2" = "blue"),
               form = "circular")
}

}
\seealso{
\code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{ggsave}},
\code{\link[ggtree]{ggtree}}, \code{\link[ggtree]{geom_hilight}}
}
