% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{node_ids}
\alias{node_ids}
\title{node_ids: Print a phylogenetic tree with node IDs and an option to print tip labels matching a user-provided pattern.
This function generates a plot of a phylogenetic tree with node IDs displayed. Additionally, it offers the option
to highlight specific tip labels that match a user-provided pattern.}
\usage{
node_ids(
  tree,
  form = "circular",
  node_id_color = "darkred",
  tip_label_size = 2,
  ...
)
}
\arguments{
\item{tree}{An object representing the phylogenetic tree. Should be of class 'treedata' or 'phylo'.}

\item{form}{Layout of the tree based on ggtree options. Layout can be rectangular, circular, roundrect, slanted, ellipse, fan, equal_angle, daylight (Default: "rectangular").}

\item{node_id_color}{Color of the printed node ids. Default is "darkred".}

\item{...}{Pattern(s) for printing tip labels that match them.}
}
\value{
A ggplot object representing the phylogenetic tree with node IDs and optional highlighted tip labels.
}
\description{
node_ids: Print a phylogenetic tree with node IDs and an option to print tip labels matching a user-provided pattern.
This function generates a plot of a phylogenetic tree with node IDs displayed. Additionally, it offers the option
to highlight specific tip labels that match a user-provided pattern.
}
\examples{
\dontrun{
# Load a sample tree
tree <- read_tree("path/to/tree.nwk")

# Print the tree with node IDs and print tip labels matching "taxon_A"
node_ids(tree, color = "darkred", "taxon_A")
}

}
