% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTRee.R
\name{print_internal_nodes}
\alias{print_internal_nodes}
\title{print_internal_nodes}
\usage{
print_internal_nodes(
  tree,
  form = "circular",
  node_id_color = "darkred",
  pattern = NULL,
  tip_label_size = 2,
  output = "node_with.svg"
)
}
\arguments{
\item{tree}{An object representing the phylogenetic tree. Should be of class 'treedata' or 'phylo'.}

\item{form}{Layout of the tree based on ggtree options. Layout can be rectangular, circular, roundrect, slanted, ellipse, fan, equal_angle, daylight. Default: "rectangular"}

\item{node_id_color}{Color of the printed node labels. Default: 'darkred'}

\item{pattern}{Pattern or list of patterns for matching tip labels of interest. If pattern is NULL, then only node labels will be printed. Default: NULL}

\item{tip_label_size}{Size of printed tip labels matching the user-specified pattern. Default: 2}
}
\value{
A ggplot object representing the phylogenetic tree with node labels and optional highlighted tip labels.
Used for identifying nodes for interest for extracting subtrees, flipping or labeling them in the visualize_tree function.
}
\description{
This function generates a plot of the provided phylogenetic tree with node labels displayed.
Optional printing of tip labels matching a user-provided pattern.
}
\examples{
\dontrun{
  # Load a sample tree
  tree <- read_tree('tree.nwk')

  # Print the tree with node labels and print tip labels matching "taxon_A"
  print_internal_nodes(tree, color = 'darkred', pattern = 'taxon_A' )
}

}
