% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTRee.R
\name{write_newick}
\alias{write_newick}
\title{write_newick}
\usage{
write_newick(tree, output = NULL)
}
\arguments{
\item{tree}{An object representing the phylogenetic tree. Should be of class 'treedata' or 'phylo', or a file.}

\item{output}{Name of the output file. If no output is specified, is provided it will use the tree object name with a '.nwk' suffix.}
}
\description{
Writes a tree in newick format, while preserving bootstrap support values (if present).
}
\examples{
\dontrun{
Read a tree in nexus format
  nx_tree <- read_tree(input_file = 'tree.nxs')
  
Write it in newick format
  write_newick( tree = nx_tree, output = 'nx_tree.nwk' )
}

}
