% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{write_tree}
\alias{write_tree}
\title{write_tree
This function writes a tree in newick format, while preserving the bootstrap values.}
\usage{
write_tree(tree, output = "output_tree.nwk")
}
\arguments{
\item{tree}{An object representing the phylogenetic tree. Should be of class 'treedata' or 'phylo', or a file.}

\item{output}{Name of the output file. If no name, is provided it will use the tree object name with a '.nwk' suffix.}
}
\value{
A ggplot object representing the phylogenetic tree with node IDs and optional highlighted tip labels.
}
\description{
write_tree
This function writes a tree in newick format, while preserving the bootstrap values.
}
