% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{bootstrap_collapse}
\alias{bootstrap_collapse}
\title{bootstrap_collapse: Collapse nodes in a phylogenetic tree based on bootstrap support.
This function collapses nodes in a phylogenetic tree where the bootstrap support is below a specified cutoff.}
\usage{
bootstrap_collapse(tree, cutoff = 0.5)
}
\arguments{
\item{tree}{An object representing the phylogenetic tree. Should be of class 'treedata' or 'phylo'.}

\item{cutoff}{The threshold value for collapsing nodes based on bootstrap support. Defaults to 0.5.}
}
\value{
A 'phylo' object with collapsed branches with bootstrap less than the provided cutoff.
}
\description{
bootstrap_collapse: Collapse nodes in a phylogenetic tree based on bootstrap support.
This function collapses nodes in a phylogenetic tree where the bootstrap support is below a specified cutoff.
}
\examples{
\dontrun{
# Load a sample tree
tree <- read_tree("path/to/tree.nwk")

# Collapse nodes with bootstrap support less than 0.7
collapsed_tree <- bootstrap_collapse(tree, 0.7)
}

}
