% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{.load_tree_object}
\alias{.load_tree_object}
\title{.load_tree_object
Load phylogenetic tree as treedata object}
\usage{
.load_tree_object(tree = NULL)
}
\arguments{
\item{tree}{An object representing the phylogenetic tree, providing a single tree entrypoint used in the whole package.
Makes use of the read_tree function.

If no file path is provided, the user will be prompted to provide an input phylogenetic tree through Rstudio's graphical user interface (GUI).}
}
\value{
A tree data object representing the phylogenetic tree.
}
\description{
.load_tree_object
Load phylogenetic tree as treedata object
}
