% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTRee.R
\name{export_plot}
\alias{export_plot}
\title{export_plot}
\usage{
export_plot(plot, output = NULL, format = "svg", dpi = 600)
}
\arguments{
\item{plot}{A ggtree plot object.}

\item{output}{Name of the output file.}

\item{format}{Format of the output file. Default: 'svg'}

\item{dpi}{Number of DPIs in output file. Default: 600}
}
\value{
A ggplot object representing the phylogenetic tree and optionally an output svg file with this phylogenetic tree.
}
\description{
Function to export plot with provided options
}
\examples{
\dontrun{
  # Export a phylogenetic tree object
  export_plot (plot, output = "tree_plot_visualized.svg")
}
}
