% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{extract_subtree}
\alias{extract_subtree}
\title{extract_subtree: Extract Subtree
Function to extract a subtree by finding the MRCA (Most Recent Common Ancestor) of two anchor nodes while preserving branch lengths and bootstrap values.}
\usage{
extract_subtree(tree, tip1, tip2)
}
\arguments{
\item{tree}{A phylogenetic tree object of class 'phylo' or 'treedata'.}

\item{tip1}{The label or pattern of the first tip node.}

\item{tip2}{The label or pattern of the second tip node.}

\item{branch_length}{Logical. If TRUE, preserve branch lengths in the extracted subtree; if FALSE, create a cladogram with equal branch lengths.}
}
\value{
A treedata object representing the extracted subtree with preserved branch lengths and bootstrap values.
}
\description{
extract_subtree: Extract Subtree
Function to extract a subtree by finding the MRCA (Most Recent Common Ancestor) of two anchor nodes while preserving branch lengths and bootstrap values.
}
\examples{
# Extract a subtree preserving branch lengths and bootstrap values
tree <- read.tree("path/to/your/treefile.newick")
subtree <- extract_subtree(tree, "species_A", "species_B", branch_length = TRUE)
visualize_tree(subtree)

# Extract a cladogram with equal branch lengths
tree <- read.tree("path/to/your/treefile.newick")
cladogram <- extract_subtree(tree, "species_A", "species_B", branch_length = FALSE)
visualize_tree(cladogram)

}
\seealso{
\code{\link{visualize_tree}} for visualizing phylogenetic trees.
\code{\link{findMRCA}} for finding the Most Recent Common Ancestor (MRCA) of multiple nodes.
}
