% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R, R/OliveTree_CMD.R
\name{highlight_tree}
\alias{highlight_tree}
\title{highlight_tree}
\usage{
highlight_tree(
  tree,
  highlight_nodes,
  colors = NULL,
  layout = "circular",
  name = NULL,
  ...
)

highlight_tree(
  tree,
  highlight_nodes,
  colors = NULL,
  layout = "circular",
  name = NULL,
  ...
)
}
\arguments{
\item{tree}{A phylogenetic tree object of class 'phylo' or a file path to a tree file (e.g., in Newick format).}

\item{highlight_nodes}{A associative vector of node labels with associated group names to highlight on the tree.}

\item{colors}{An associative vector of colors with optional group names to connect with the highlight nodes. If not provided, random colors will be assigned.}

\item{layout}{A character string specifying the layout of the tree. Options include "rectangular", "circular", etc.}

\item{name}{A character string specifying the name for the highlighted plot.}
}
\value{
A ggtree object with highlighted nodes.

A ggtree object with highlighted nodes.
}
\description{
Function to highlight nodes and descendant leaves on a phylogenetic tree.

Function to highlight nodes and descendant leaves on a phylogenetic tree.
}
\examples{
# Highlight specific nodes with random colors
tree <- read.tree("path/to/your/treefile.newick")
highlight_tree(tree, c("species_A", "species_B"))

# Highlight specific nodes with specified colors
tree <- read.tree("path/to/your/treefile.newick")
highlight_tree(tree, 
               highlight_nodes = c("Category_A" = node_number_A, 
                                   "Category_B" = node_number_B), 
               colors = c("Category_A" = color_A, 
                          "Category_B" = color_B)
               )

# Highlight specific nodes with random colors
tree <- read.tree("path/to/your/treefile.newick")
highlight_tree(tree, c("species_A", "species_B"))

# Highlight specific nodes with specified colors
tree <- read.tree("path/to/your/treefile.newick")
highlight_tree(tree, c("species_A", "species_B"), colors = c("red", "blue"))

}
\seealso{
\code{\link{visualize_tree}} for visualizing phylogenetic trees.

\code{\link{visualize_tree}} for visualizing phylogenetic trees.
}
