% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R, R/plotter.R
\name{load_packages}
\alias{load_packages}
\title{load_packages: Function to check if a package is installed, and if not, install it.}
\usage{
load_packages(tools)

load_packages(tools)
}
\arguments{
\item{tools}{A character vector of package names to be checked and installed.}
}
\description{
This function checks if a package is installed. If not, it installs the package using BiocManager if available, otherwise using install.packages.

This function checks if a package is installed. If not, it installs the package using BiocManager if available, otherwise using install.packages.
}
\details{
Library of functions for advanced tree manipulation and visualization using ggtree, ape, phytools, and other related tools.

Library of plotting functions.
}
