% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{print_internal_nodes}
\alias{print_internal_nodes}
\title{print_internal_nodes: This function generates a plot of the provided phylogenetic tree with node IDs displayed.
Additionally, it offers the option to highlight specific tip labels that match a user-provided pattern.}
\usage{
print_internal_nodes(
  tree,
  form = "circular",
  node_id_color = "red",
  tip_label_size = 2,
  references = NULL,
  save = TRUE,
  output = "highlighted_tree.svg"
)
}
\arguments{
\item{tree}{An object representing the phylogenetic tree. Should be of class 'treedata' or 'phylo'.}

\item{form}{Layout of the tree based on ggtree options. Layout can be rectangular, circular, roundrect, slanted, ellipse, fan, equal_angle, daylight (Default: "rectangular").}

\item{node_id_color}{Color of the printed node ids. Default is "darkred".}

\item{...}{Pattern(s) for printing tip labels that match them.}
}
\value{
A ggplot object representing the phylogenetic tree with node IDs and optional highlighted tip labels.
Used for identifying nodes for interest for extracting subtrees, flipping or labeling them in the visualize_tree function.
}
\description{
print_internal_nodes: This function generates a plot of the provided phylogenetic tree with node IDs displayed.
Additionally, it offers the option to highlight specific tip labels that match a user-provided pattern.
}
\examples{
\dontrun{
# Load a sample tree
tree <- read_tree("path/to/tree.nwk")

# Print the tree with node IDs and print tip labels matching "taxon_A"
print_internal_nodes(tree, color = "darkred", "taxon_A")
}

}
