% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{read_tree}
\alias{read_tree}
\title{read_tree: Read a phylogenetic tree.
This function reads a phylogenetic tree from a file.}
\usage{
read_tree(input_file, bootstrap_support = TRUE)
}
\arguments{
\item{input_file}{Path to the file containing the phylogenetic tree.}
}
\value{
An object representing the phylogenetic tree.
}
\description{
read_tree: Read a phylogenetic tree.
This function reads a phylogenetic tree from a file.
}
\examples{
\dontrun{
tree <- read_tree("path/to/tree.nwk")
}

}
