% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{visualize_tree}
\alias{visualize_tree}
\title{visualize_tree: Visualize Phylogenetic Tree with Various Customizable Features}
\usage{
visualize_tree(
  tree = NULL,
  form = "rectangular",
  flip_nodes = FALSE,
  node1,
  node2 = NULL,
  tiplabels = FALSE,
  pattern_id = NULL,
  color = NULL,
  shape = NULL,
  taxon_group_separator = "NULL",
  legend_position = "right",
  legend_text_position = "right",
  legend_orientation = "horizontal",
  legend_key_size = 0.1,
  legend_font = "Arial",
  legend_fontsize = 11,
  legend_font_face = "bold",
  legend_spacing_x = 0.5,
  legend_spacing_y = 0.5,
  legend_key_width = 1,
  legend_title_hjust = 0.5,
  legend_name = "",
  tip_shape_size = 3,
  tip_label_size = 2,
  tip_label_color = "black",
  bootstrap_numbers = FALSE,
  bootstrap_number_nudge_x = 0,
  bootstrap_number_nudge_y = 0.2,
  node_label_size = 3,
  bootstrap_circles = FALSE,
  bootstrap_legend = FALSE,
  bootstrap_circle_size = 1.7,
  branch_length = TRUE,
  clades = NULL,
  labeldist = 1,
  bardist = 1,
  clade_label_size = 3,
  save = TRUE,
  output = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree in Newick format, treedata, or phylo class object.}

\item{form}{The layout of the tree, with options including "rectangular", "circular",
"roundrect", "slanted", "ellipse", "fan", "equal_angle", and "daylight". Default is "rectangular".}

\item{flip_nodes}{Logical; whether to flip the tree at specified nodes.}

\item{node1, node2}{The ID of the two nodes to flip based on their most common recent ancestor (MRCA)}

\item{tiplabels}{Logical; whether to print tip labels on the tree. Default: FALSE.}

\item{pattern_id}{A pattern used to match and selectively print tip labels as text.}

\item{color}{A vector specifying color mappings for tip label shapes based on the taxon ID.
Is different to tip_label_color, to allow for a user to provide colors for both text and shape mappings.}

\item{shape}{A vector specifying ggstar starshapes for tip labels based on the taxon ID.}

\item{taxon_group_separator}{A separator string for splitting taxon IDs to generate color
and shape mappings. If NULL, mappings are generated based on the whole taxon ID string.}

\item{legend_position}{The position of the legend on the plot ("right" by default).}

\item{legend_text_position}{The position of the text in the legend ("right" by default).}

\item{legend_orientation}{The orientation of the legend, either "horizontal" or "vertical" (default).}

\item{legend_key_size}{Size of the keys in the legend, in cm.}

\item{legend_font}{The font family for the legend text.}

\item{legend_fontsize}{The size of the font in the legend, in points.}

\item{legend_font_face}{Whether the font in the legend should be bold ("bold" or normal).}

\item{legend_spacing_x}{Horizontal spacing in the legend, in cm.}

\item{legend_spacing_y}{Vertical spacing in the legend, in cm.}

\item{legend_key_width}{The width of the legend keys, in cm.}

\item{legend_title_hjust}{Horizontal justification of the legend title (0 to 1).}

\item{legend_name}{The name to be displayed as the legend title.}

\item{tip_shape_size}{Numeric specifying the size of the ggstar starshapes representing tip labels.}

\item{tip_label_size}{Numeric specifying the size of the tip labels.}

\item{tip_label_color}{Color for the tip labels, default is black.}

\item{bootstrap_numbers}{Logical; whether to display bootstrap values on branches.}

\item{bootstrap_number_nudge_y}{Numeric value controlling the vertical offset of bootstrap numbers.}

\item{bootstrap_circles}{Logical; whether to display bootstrap values as circles at nodes.}

\item{bootstrap_legend}{Logical; whether to display a legend for bootstrap values if displayed as circles.}

\item{bootstrap_circle_size}{Numeric specifying the size of the bootstrap circles.}

\item{branch_length}{Logical indicating whether branch lengths should be shown or omitted. Default: TRUE}

\item{clades}{A vector or list specifying clades and their labels for highlighting.}

\item{labeldist}{The distance for clade labels from their respective nodes.}

\item{bardist}{The distance for clade label bars from their respective nodes.}

\item{clade_label_size}{The size of the clade labels.}

\item{save}{Logical; if TRUE, the plot will be saved using the specified output file name.}

\item{output}{The file name or path where the plot will be saved if \code{save} is TRUE.}
}
\value{
A ggplot object representing the visualized tree.
}
\description{
Main tree visualization function of the package, which provides comprehensive options
for visualizing a phylogenetic tree, allowing for customizations such as
displaying bootstrap values, tip labels as text and shape with different colors based on user-provided mappings,
flipping nodes and label user-specified clades. It supports various tree layouts.
}
\examples{
\dontrun{
library(ggtree)
library(ggplot2)
tree <- rtree(50)  # Generate a random tree
visualize_tree(tree, form = "circular", bootstrap_circles = TRUE,
               color = c("Species1" = "red", "Species2" = "blue", "Species3" = "orange"),
               shape = c("Species1" = "circle", "Species2" = "star", "Species3" = "triangle"),
               tiplabels = TRUE, pattern_id = "Species")
}

}
\seealso{
\link[ggtree]{ggtree}, \link[ggplot2]{geom_point}
}
