% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OliveTree.R
\name{visualize_tree}
\alias{visualize_tree}
\title{Function to visualize a tree with a wide variety of options and customizable features
Reference taxons should be defined as reference = 'taxon_ID'.}
\usage{
visualize_tree(
  tree,
  form = "rectangular",
  tiplabels = FALSE,
  pattern_id = NULL,
  color = NULL,
  shape = NULL,
  mappings_legend = FALSE,
  tip_shape_size = 3,
  tip_label_size = 2,
  fontsize = 3,
  bootstrap_numbers = TRUE,
  bootstrap_number_nudge_x = 1,
  bootstrap_number_nudge_y = 3.2,
  node_label_size = 3,
  bootstrap_circles = FALSE,
  bootstrap_legend = FALSE,
  branch_length = TRUE,
  clades = NULL,
  labeldist = 1,
  bardist = 1,
  save = TRUE,
  output = NULL,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{tree}{Phylogenetic tree in Newick format, treedata, or phylo class object.}

\item{form}{Layout of the tree based on ggtree options. Layout can be rectangular, circular, roundrect, slanted, ellipse, fan, equal_angle, daylight (Default: "rectangular").}

\item{tiplabels}{Logical. Print tip labels on the tree.}

\item{pattern_id}{Pattern to match tip labels for printing.}

\item{color}{Vector specifying tip color mappings.}

\item{shape}{Vector specifying tip shape mappings.}

\item{mappings_legend}{Logical. Display legend for color and shape mappings.}

\item{tip_shape_size}{Numeric. Size of tip shapes.}

\item{tip_label_size}{Numeric. Size of tip labels.}

\item{bootstrap_numbers}{Logical. Display bootstrap values on branches. Default: TRUE}

\item{bootstrap_number_nudge_y}{Numeric. Controls the relative height of the bootstrap number on top of the branch.}

\item{bootstrap_circles}{Logical. Display bootstrap values as circles on parent nodes of branches.}

\item{bootstrap_legend}{Logical. Display legend for bootstrap circles.}

\item{clades}{Associative vector of node IDs and related labels.}

\item{save}{Logical. Save the plot.}

\item{output}{Character. Output file name if saving the plot.}

\item{interactive}{Logical. Generate an interactive plot using plotly.}

\item{...}{Add pattern(s) for reference taxon IDs. Works only if color and/or shape mappings are provided}
}
\value{
A ggplot or ggplotly object representing the visualized tree.
}
\description{
Function to visualize a tree with a wide variety of options and customizable features
Reference taxons should be defined as reference = 'taxon_ID'.
}
\examples{
\dontrun{
# Load required libraries
library(ggplot2)
library(ggtree)

# Example: Visualize a tree with bootstrap values as circles and custom colors
visualize_tree(tree, form = "rectangular", bootstrap_circles = TRUE,
               color = c(species1 = "red", species2 = "blue"))
}

}
\seealso{
\code{\link[ggtree]{ggtree}}, \code{\link[ggtree]{geom_tiplab}},
\code{\link[ggtree]{geom_nodelab}}, \code{\link[ggtree]{geom_cladelab}},
\code{\link[ggstar]{geom_star}}, \code{\link[ggplot2]{theme_tree}}, \code{\link[ggplot2]{ggsave}},
\code{\link[plotly]{ggplotly}}
}
